+++ VaccineC/VaccineC.Command.Application/Commands/ProductSummaryBatch/DeleteProductSummaryBatchCommandHandler.cs
@@ -15,6 +15,12 @@ namespace VaccineC.Command.Application.Commands.ProductSummaryBatch
         public async Task<Unit> Handle(DeleteProductSummaryBatchCommand request, CancellationToken cancellationToken)
         {
             var productSummaryBatch = _repository.GetById(request.Id);
+
+            if (productSummaryBatch == null)
+            {
+                throw new ArgumentException("Produto Lote não encontrado!");
+            }
+
             _repository.Remove(productSummaryBatch);
             await _repository.SaveChangesAsync();
