+++ VaccineC/VaccineC.Command.Application/Commands/UserCommands/DeleteUserCommandHandler.cs
@@ -15,6 +15,12 @@ namespace VaccineC.Command.Application.Commands.UserCommands
         public async Task<Unit> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
         {
             var user = _userRepository.GetById(request.Id);
+
+            if (user == null)
+            {
+                throw new ArgumentException("Usuário não encontrado!");
+            }
+
             _userRepository.Remove(user);
             await _userRepository.SaveChangesAsync();
